fix upload dan database

+++ app/models/Supplier_model.php
@@ -117,29 +117,33 @@ class Supplier_model
 
 	public function uploadFile($file, $type = [], $targetDir = 'upload/', $maxSize = 2*MB, $oldFileName = '')
     {
-		if (!empty($oldFileName)) 
-			$this->deleteFile($targetDir . '/' . $oldFileName);
+		if ($file['error'] !== 4) {
+			if (!empty($oldFileName)) 
+				$this->deleteFile($targetDir . '/' . $oldFileName);
 
-        $name = $file['name'];
+			$name = $file['name'];
 
-		if ($file["size"] > $maxSize)
-            return false;
-        
-        $imageFileType = explode('.', $name);
-        $imageFileType = strtolower(end($imageFileType));
-        if (!in_array($imageFileType, $type))
-            return false;
+			if ($file["size"] > $maxSize)
+				return false;
+			
+			$imageFileType = explode('.', $name);
+			$imageFileType = strtolower(end($imageFileType));
+			if (!in_array($imageFileType, $type))
+				return false;
 
-        $fileName = uniqid() . "." . $imageFileType;
-        $targetFile = $targetDir . $fileName;
+			$fileName = uniqid() . "." . $imageFileType;
+			$targetDir .= $fileName;
 
-        try {
-            move_uploaded_file($file['tmp_name'], $targetFile);
-        } catch (Exception $e) {
-            echo $e; die;
-        }
+			try {
+				move_uploaded_file($file['tmp_name'], $targetDir);
+			} catch (Exception $e) {
+				echo $e; die;
+			}
 
-        return $fileName;
+			return $fileName;
+		} else {
+			return empty($oldFileName) ? false : $oldFileName;
+		}
     }
 
     public function deleteFile($filepath)
