+++ app/models/Template_model.php
@@ -18,7 +18,7 @@ class Template_model
 	public function getAllData()
 	{
 		$this->db->query("SELECT * FROM {$this->table} WHERE `status` = 1");
-	return $this->db->fetchAll();
+		return $this->db->fetchAll();
 	}
 
 	public function getDataById($id)
@@ -60,6 +60,8 @@ class Template_model
 			WHERE id = :id"
 		);
 
+		$old = $this->getDataById($id);
+
 		foreach ($this->fields as $field) $this->db->bind($field, $data[$field]);
 		$this->db->bind('id', $id);
 		$this->db->bind('modified_by', $this->user);
@@ -100,29 +102,33 @@ class Template_model
 
 	public function uploadFile($file, $type = [], $targetDir = 'upload/', $maxSize = 2*MB, $oldFileName = '')
     {
-		if (!empty($oldFileName)) 
-			$this->deleteFile($targetDir . '/' . $oldFileName);
+		if ($file['error'] !== 4) {
+			if (!empty($oldFileName)) 
+				$this->deleteFile($targetDir . '/' . $oldFileName);
 
-        $name = $file['name'];
+			$name = $file['name'];
 
-		if ($file["size"] > $maxSize)
-            return false;
-        
-        $imageFileType = explode('.', $name);
-        $imageFileType = strtolower(end($imageFileType));
-        if (!in_array($imageFileType, $type))
-            return false;
+			if ($file["size"] > $maxSize)
+				return false;
+			
+			$imageFileType = explode('.', $name);
+			$imageFileType = strtolower(end($imageFileType));
+			if (!in_array($imageFileType, $type))
+				return false;
 
-        $fileName = uniqid() . "." . $imageFileType;
-        $targetDir .= $fileName;
+			$fileName = uniqid() . "." . $imageFileType;
+			$targetDir .= $fileName;
 
-        try {
-            move_uploaded_file($file['tmp_name'], $targetDir);
-        } catch (Exception $e) {
-            echo $e; die;
-        }
+			try {
+				move_uploaded_file($file['tmp_name'], $targetDir);
+			} catch (Exception $e) {
+				echo $e; die;
+			}
 
-        return $fileName;
+			return $fileName;
+		} else {
+			return empty($oldFileName) ? false : $oldFileName;
+		}
     }
 
     public function deleteFile($filepath)
